Move fake auth dialog
